pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot_framework"
version = "0.0.1"
authors = [
  { name="ITK Development", email="itk-rpa@mkb.aarhus.dk" },
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "OpenOrchestrator == 1.*",
    "Pillow == 10.*",
    "requests == 2.32.3",
    "office365-rest-python-client==2.5.14"
]

[project.optional-dependencies]
dev = [
  "pylint",
  "flake8"
]
